Article: "Introducing stb-tester"
Note that the video is not ready yet; the thumbnails are just placeholders.

The CSS changes are thus:

The new article uses a very simple (fixed-width, centered) layout. Only the
index page uses the robot-hand-with-remote-control image, content at a fixed
(left margin) distance from the image, varying the image and margin size as
the window width changes.

There are also some minor corrections to top & bottom margins of the header,
content, and footer divs such that the margins are a multiple of the text's
line height.