"make deploy": Copies html to new commit on gh-pages branch
We host the stb-tester.com website on "GitHub pages". See
https://help.github.com/articles/creating-project-pages-manually

"make deploy" commits the generated html to a new commit on the gh-pages
branch. Doesn't actually push the branch to origin so that you can
review the changes first. Remember to switch back to the master branch
when you're done!