New article: Using Jenkins to schedule & report stb-tester scripts
The "detailed instructions" section isn't done yet, but the "system
overview" is (hopefully) useful enough that it merits being published
now. When I publish the "detailed instructions", I must remember to
update the <updated> node of the atom feed, so that it will be
highlighted in people's readers.

On the index page, the link is added to a separate <ul> under the
Documentation section, to separate the "core" documentation from
more bloggy articles that may not be relevant to everyone.

Re. the css changes:

* Class `figure` still looks the same in the existing articles, but now
  has better default margins.
* Header h1 has a smaller, more sensible size -- at 36pt this article's
  title was wrapping to 3 lines. Only the home page and the man page
  keep the 36pt header: it actually looks good there, and they have
  short titles.
* The rules `#jenkins a[href^="http://stb-tester.com"]` and
  `#jenkins a[href^="http://jenkins-ci.org"]` are necessary because
  reStructuredText doesn't allow me to nest inline markup like
  `**stb-tester**`_ or **`stb-tester`_** for bold links.