CSS of man page: Consistent margins in API documentation
This affects the top & bottom margins of p, ul and dd elements in the
"Test script format" section documenting each stbt function.

Technically it affects elements in the whole document, but the
difference is only noticeable in the API documentation, where the block
elements are immediately preceded by a dt element with no bottom margin.

Note that elements with only one paragraph of docstring only end up with
<dd>...</dd>, not <dd><p>...</p></dd>. (That is the way that rst2html
generates the html, so I don't think I can control it.)