Getting started: Installation instructions for RedHat, Debian & OS X
On each of these systems the package names are different, so document
the differences (in footnotes, so the prose doesn't become overloaded
with technical details).

We now offer pre-built RPMs for Fedora, so I also added the relevant
instructions.

A few other minor changes:
* Explain videotestsrc and ximagesink.
* Mention udp{sink,src} instead of tcp{client,server}{sink,src} because
we've actually used the former successfully.