Makefile: Compatibility fix for GNU make 3.82
With GNU make 3.81, this Makefile was working fine. With 3.82, I was
getting this error:

    sed: 1: "/<start python docs>/,/ ...": invalid command code \

The backslash at the end of the line was being seen literally by sed
(because it was inside single quotes). Presumably the older version of
make was stripping the backslash out completely. The new version of make
behaves as per the specification:
http://www.gnu.org/software/make/manual/html_node/Splitting-Lines.html

The solution here is to use double quotes instead of single quotes.
Consider these examples of shell behaviour:

    $ echo 'a\
    > b'
    a\
    b

    $ echo "a\
    > b"
    ab