match-parameters: layout/alignment change
This commit makes the figures align to the right hand side of the
containers, along with their labels. This means that the body text can
happily align left and wrap around the images which makes for easier
reading. The increase to .clear .figure margin makes the figures sit
slightly lower so that visually the text comes first.

The additional clear container was required in the article to prevent
the literal text block (the command) from overlaying on of the figures.